requirements.txt
Django>=4.2.0,<4.3.0
markdown>=3.4.3,<3.5.0
psycopg[binary]>=3.1.8,<3.2.0

django-filter>=23.1.0,<23.2.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.2.2,<5.3.0
django-cors-headers>=3.14.0,<3.15.0
#djangorestframework-jsonapi>=6.0.0,<6.1.0

flake8>=6.0.0,<6.1.0
black>=23.3.0,<23.4.0
isort>=5.12.0,<5.13.0
mypy>=1.2.0,<1.3.0
django-stubs>=1.16.0,<1.17.0
djangorestframework-stubs>=1.10.0,<1.11.0
